+++ main/src/main/kotlin/de/henningwobken/vpex/main/model/Find.kt
@@ -1,3 +1,3 @@
 package de.henningwobken.vpex.main.model
 
-data class Find(val start: Int, val end: Int)
+data class Find(val start: Long, val end: Long)
